AjaxJSONSampleCode
Here is sample code to load server files in HTML, txt, JSON, and XML.  I provides examples of loading, parsing, accessing and displaying content from different file formats and content.  It also show how to set properties based on information found on the server.  This pattern can be used to access internalStorage to manage and control the look and feel.